requirements.txt
Flask
Pillow==10.1.0
